+++ tests/team/test_client.py
@@ -37,3 +37,18 @@ def test_get_team_by_name(requests_mock):
         status_code=200,
     )
     assert TeamClient().get_team_by_name(team_name) == team
+
+
+def test_get_team_members(requests_mock):
+    team_id = "abc"
+    members = [
+        {"name": "Foo", "username": "foo"},
+        {"name": "Bar", "username": "bar"},
+    ]
+    requests_mock.register_uri(
+        "GET",
+        re.compile(f"teams/{team_id}/members"),
+        text=json.dumps(members),
+        status_code=200,
+    )
+    assert TeamClient().get_team_members(team_id) == members